01_ai_agents_first/26_external_tracing_and_basic_evals/basic_eval/pyproject.toml
[project]
name = "basic-eval"
version = "0.2.0"
description = "Basic agent evaluation with Langfuse and OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langfuse>=3.6.1",
    "openai-agents>=0.3.3",
    "openinference-instrumentation-openai-agents>=1.3.0",
    "python-dotenv>=1.1.1",
]
